move monitor image routime back into api boot; fix to use AppDockerLogin
